refactor(relay): move the frame parsers out of the component file
parseRelayFrameSet and parseRelayText are pure and DOM-free. Their tests were
under tests/ui/ only because the functions sat in a .tsx file.
